backend/payments/serializers/account.py
from rest_framework import serializers

from payments.models import Account


class AccountSerializer(serializers.ModelSerializer):
    class Meta:
        model = Account
        fields = ('id', 'balance')
