backend/requirements.txt
Django<=3.0
psycopg2-binary<=2.9
python-dotenv==0.10.2
djangorestframework<=3.10
django-cors-headers<3.1
django-money==0.15
bcrypt<3.2
django-model-utils<3.2
